+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/data/model/Mode.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import org.springframework.lang.Nullable;
 
 import java.util.EnumSet;
+import java.util.Objects;
 
 @Getter
 public enum Mode {
@@ -14,7 +15,7 @@ public enum Mode {
 	private final String modeName;
 	private final String folder;
 
-	Mode(String modeName, String folder) {
+	Mode(@Nullable String modeName, String folder) {
 		this.modeName = modeName;
 		this.folder = folder;
 	}
@@ -22,6 +23,8 @@ public enum Mode {
 	public static Mode getEnumFromModeName(String modeName) {
 		return EnumSet.allOf(Mode.class)
 				.stream()
+				.filter(Objects::nonNull)
+				.filter(mode->mode.getModeName()!=null)
 				.filter(mode -> mode.getModeName().equals(modeName))
 				.findFirst()
 				.orElseThrow(() -> new IllegalStateException(String.format("Unsupported mode %s.", modeName)));