+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/exceptions/NullException.java
@@ -1,7 +1,10 @@
 package fr.insee.kraftwerk.core.exceptions;
 
+import java.io.Serial;
+
 public class NullException extends KraftwerkException {
 
+    @Serial
     private static final long serialVersionUID = 6677487610288558193L;