+++ kraftwerk-core/src/test/java/cucumber/unit_tests/EvalScriptDefinitions.java
@@ -1,7 +1,5 @@
 package cucumber.unit_tests;
 
-import static org.junit.Assert.assertEquals;
-
 import fr.insee.kraftwerk.core.rawdata.SurveyRawData;
 import fr.insee.kraftwerk.core.rawdata.SurveyRawDataTest;
 import fr.insee.kraftwerk.core.utils.files.FileSystemImpl;
@@ -12,6 +10,8 @@ import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 
+import static org.junit.Assert.assertEquals;
+
 // Used in do_we_apply_vtl_instruction and do_we_apply_vtl_script
 public class EvalScriptDefinitions {
 	public VtlBindings vtlBindings = new VtlBindings();
@@ -20,7 +20,7 @@ public class EvalScriptDefinitions {
 	KraftwerkExecutionContext kraftwerkExecutionContext = new KraftwerkExecutionContext();
 
 	@Given("We have some simple VTLBindings")
-	public void initialize() throws Exception {
+	public void initialize() {
 		SurveyRawData surveyColeman = SurveyRawDataTest.createFakeCawiSurveyRawData();
 		vtlExecute.convertToVtlDataset(surveyColeman, "COLEMAN", vtlBindings);
 		SurveyRawData surveyPaper = SurveyRawDataTest.createFakePapiSurveyRawData();
@@ -30,7 +30,7 @@ public class EvalScriptDefinitions {
 	} 
 	
 	@When("I try to apply some VTL instruction : {string}")
-	public void exportDataset(String vtlScript) throws Exception {
+	public void exportDataset(String vtlScript) {
 		vtlExecute.evalVtlScript(vtlScript, vtlBindings, kraftwerkExecutionContext);
 	}