+++ kraftwerk-core/src/test/java/fr/insee/kraftwerk/core/dataprocessing/CalculatedProcessingTest.java
@@ -19,7 +19,10 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class CalculatedProcessingTest {
 
@@ -67,12 +70,12 @@ class CalculatedProcessingTest {
     @Test
     void testIfCalculatedAreProcessed() {
         //
-        VtlBindings vtlBindings = getVtlBindings();
+        VtlBindings vtlBindingsTest = getVtlBindings();
         //
-        CalculatedProcessing processing = new CalculatedProcessing(vtlBindings, fooCalculated, fileUtilsInterface);
+        CalculatedProcessing processing = new CalculatedProcessing(vtlBindingsTest, fooCalculated, fileUtilsInterface);
         processing.applyAutomatedVtlInstructions("TEST", kraftwerkExecutionContext);
         //
-        Dataset outDataset = vtlBindings.getDataset("TEST");
+        Dataset outDataset = vtlBindingsTest.getDataset("TEST");
 
         //
         assertNotNull(outDataset);
@@ -81,9 +84,9 @@ class CalculatedProcessingTest {
         assertTrue(outDataset.getDataStructure().containsKey("FOO2"));
         assertTrue(outDataset.getDataStructure().containsKey("FOO3"));
         //
-        assertEquals(1L, outDataset.getDataPoints().get(0).get("FOO3"));
-        assertEquals(1L, outDataset.getDataPoints().get(0).get("FOO2"));
-        assertEquals(2L, outDataset.getDataPoints().get(0).get("FOO1"));
+        assertEquals(1L, outDataset.getDataPoints().getFirst().get("FOO3"));
+        assertEquals(1L, outDataset.getDataPoints().getFirst().get("FOO2"));
+        assertEquals(2L, outDataset.getDataPoints().getFirst().get("FOO1"));
     }
 
     @NotNull