+++ kraftwerk-core/src/test/java/fr/insee/kraftwerk/core/vtl/TrevasTests.java
@@ -7,16 +7,23 @@ import fr.insee.vtl.model.InMemoryDataset;
 import fr.insee.vtl.model.Structured;
 import org.junit.jupiter.api.Test;
 
-import javax.script.*;
+import javax.script.Bindings;
+import javax.script.ScriptContext;
+import javax.script.ScriptEngine;
+import javax.script.ScriptEngineManager;
+import javax.script.ScriptException;
+import javax.script.SimpleBindings;
 import java.io.IOException;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class TrevasTests {
 
     @Test
-    void largeExpressionWithNoResult() throws IOException, ScriptException {
+    void largeExpressionWithNoResult() throws IOException {
         //
         Bindings bindings = new SimpleBindings();
         ObjectMapper mapper = new ObjectMapper();
@@ -51,7 +58,7 @@ class TrevasTests {
     }
 
     @Test
-    void expressionFailingForEveryone() throws IOException, ScriptException {
+    void expressionFailingForEveryone() throws ScriptException {
         //
         Bindings bindings = new SimpleBindings();
         Dataset dataset = new InMemoryDataset(
@@ -79,7 +86,7 @@ class TrevasTests {
     }
 
     @Test
-    void elseWithSameVariable() throws IOException, ScriptException {
+    void elseWithSameVariable() throws ScriptException {
         //
         Bindings bindings = new SimpleBindings();
         Dataset dataset = new InMemoryDataset(